Move StyleComputer to nu_color_config
