+++ crates/nu-color-config/Cargo.toml
@@ -8,7 +8,14 @@ name = "nu-color-config"
 version = "0.72.2"
 
 [dependencies]
+serde = { version="1.0.123", features=["derive"] }
+# used only for text_style Alignments
+tabled = { version = "0.10.0", features = ["color"], default-features = false }
+
 nu-protocol = { path = "../nu-protocol", version = "0.72.2"  }
 nu-ansi-term = "0.46.0"
-nu-json = { path = "../nu-json", version = "0.72.2"  }
-serde = { version="1.0.123", features=["derive"] }
+nu-utils = { path = "../nu-utils", version = "0.72.2" }
+nu-engine = { path = "../nu-engine", version = "0.72.2" }
+nu-test-support = { path="../nu-test-support", version = "0.72.2"  }
+# nu-path is used only for test support
+nu-path = { path="../nu-path", version = "0.72.2"  }
