+++ crates/nu-explore/src/nu_common/mod.rs
@@ -5,8 +5,8 @@ mod value;
 
 use std::sync::{atomic::AtomicBool, Arc};
 
+use nu_color_config::TextStyle;
 use nu_protocol::Value;
-use nu_table::TextStyle;
 
 pub use nu_ansi_term::{Color as NuColor, Style as NuStyle};
 pub use nu_protocol::{Config as NuConfig, Span as NuSpan};
