+++ crates/nu-table/src/lib.rs
@@ -1,13 +1,9 @@
 mod nu_protocol_table;
-mod style_computer;
 mod table;
 mod table_theme;
-mod textstyle;
 mod util;
 
 pub use nu_protocol_table::NuTable;
 pub use table::{Alignments, Table, TableConfig};
-pub use style_computer::StyleComputer;
 pub use table_theme::TableTheme;
-pub use textstyle::{Alignment, TextStyle};
 pub use util::*;
