+++ crates/nu-table/tests/expand.rs
@@ -1,3 +1,4 @@
+use nu_color_config::{StyleComputer, TextStyle};
 mod common;
 
 use common::{create_row, create_table};
@@ -5,7 +6,7 @@ use nu_protocol::{
     engine::{EngineState, Stack},
     Config,
 };
-use nu_table::{Alignments, StyleComputer, Table, TableConfig, TableTheme as theme, TextStyle};
+use nu_table::{Alignments, Table, TableConfig, TableTheme as theme};
 
 #[test]
 fn test_expand() {
