+++ crates/nu-table/tests/style.rs
@@ -1,10 +1,11 @@
 mod common;
 
+use nu_color_config::{StyleComputer, TextStyle};
 use nu_protocol::{
     engine::{EngineState, Stack},
     Config,
 };
-use nu_table::{Alignments, StyleComputer, Table, TableConfig, TableTheme as theme, TextStyle};
+use nu_table::{Alignments, Table, TableConfig, TableTheme as theme};
 use tabled::papergrid::records::{cell_info::CellInfo, tcell::TCell};
 
 #[test]
